pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ocr-arena"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "huggingface-hub>=1.4.1",
    "httpx>=0.28.1",
    "modelscope>=1.34.0",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "portalocker>=3.2.0",
    "pytest>=8.4.2",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.20",
    "pypdfium2>=5.4.0",
    "pyyaml>=6.0.3",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.1.0",
    "uvicorn>=0.35.0",
    "wordfreq>=3.1.1",
]

[project.scripts]
ocr-arena = "ocr_arena.cli.run_local:main"
ocr-arena-layout = "ocr_arena.cli.run_layout_only:main"
ocr-arena-web = "ocr_arena.cli.run_web:main"

[[tool.uv.index]]
url = "https://mirrors.cloud.tencent.com/pypi/simple/"
default = true

[tool.setuptools.packages.find]
include = ["ocr_arena*"]
